+++ mobile/src/db/migrations/migrations.js
@@ -4,6 +4,7 @@ import m0000 from "./0000_handy_nuke.sql"
 import m0001 from "./0001_loving_bug.sql"
 import m0002 from "./0002_first_mauler.sql"
 import m0003 from "./0003_soft_iron_fist.sql"
+import m0004 from "./0004_chubby_shard.sql"
 import journal from "./meta/_journal.json"
 
 export default {
@@ -13,5 +14,6 @@ export default {
     m0001,
     m0002,
     m0003,
+    m0004,
   },
 }